% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e00-convert.R
\name{e00_convert}
\alias{e00_convert}
\title{Convert an e00 file to a binary coverage}
\usage{
e00_convert(e00, coverage = e00_temp_coverage(), coverage_type = c("V7", "PC"))
}
\arguments{
\item{e00}{Path to a .e00 file.}

\item{coverage}{An directory name whose name is less than 13 characters.
long.}

\item{coverage_type}{One of "V7" or "PC".}
}
\value{
The value of \code{coverage}.
}
\description{
Convert an e00 file to a binary coverage
}
\examples{
e00 <- system.file("e00/cis_20170911.e00", package = "arce00")
coverage <- e00_convert(e00)
sf::st_read(coverage, layer = "LAB")
unlink(coverage, recursive = TRUE)

}
