% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e00-temp.R
\name{e00_temp_coverage}
\alias{e00_temp_coverage}
\title{Create a temporary coverage directory}
\usage{
e00_temp_coverage()
}
\value{
The path to a suitable temporary coverage.
}
\description{
Binary coverages such as those created by \code{\link[=e00_convert]{e00_convert()}}
must be empty directories with names of 13 characters or fewer.
This function generates a suitable temporary directory fitting
these constraints.
}
\examples{
e00_temp_coverage()

}
