% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arce00-package.R
\docType{package}
\name{arce00-package}
\alias{arce00}
\alias{arce00-package}
\title{arce00: Read E00 Files}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/arce00}
  \item Report bugs at \url{https://github.com/paleolimbot/arce00/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey.dunnington@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
